2004-11-22	Joel Sherrill <joel@OARcorp.com>
	* score/include/rtems/score/object.h, score/src/objectclearname.c,
	score/src/objectcopynamestring.c: Fixing warning led to making
	changes to multiple prototypes for object support routines.

+++ cpukit/score/src/objectclearname.c
@@ -34,7 +34,7 @@
 
 void _Objects_Clear_name(
   void       *name,
-  uint32_t    length
+  uint16_t    length
 )
 {
   uint32_t    index;
